setting dynamic

+++ database/migrations/2023_09_13_180751_create_course_instructors_table.php
@@ -21,7 +21,7 @@ class CreateCourseInstructorsTable extends Migration
             $table->string("full_name", 100)->nullable();
             $table->string("designation", 100)->nullable();
             $table->text("details")->nullable();
-            $table->tinyInteger("creator")->unsigned()->nullable();
+            $table->bigInteger("creator")->unsigned()->nullable();
             $table->string("slug", 50)->nullable();
             $table->enum('status',['active','inactive'])->default('active');
             $table->timestamps();
