+++ database/migrations/2023_09_21_103631_create_blogs_categories_table.php
@@ -18,7 +18,7 @@ class CreateBlogsCategoriesTable extends Migration
             $table->string('title', 150)->nullable();
             $table->string('image', 100)->nullable();
 
-            $table->tinyInteger("creator")->unsigned()->nullable();
+            $table->bigInteger("creator")->unsigned()->nullable();
             $table->string("slug", 50)->nullable();
             $table->enum('status',['active','inactive'])->default('active');
